% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-tree.R
\name{as.tree}
\alias{as.tree}
\title{Convert a data frame to a tree object}
\usage{
as.tree(df, id.col = "id", parent.id.col = "parent.id", name.col = "name")
}
\arguments{
\item{df}{A data frame.}

\item{id.col}{Column name which contains the node IDs (default: "id").}

\item{parent.id.col}{Column name which contains the parent node IDs (default: "parent.id").}

\item{name.col}{Column name which contains the node names (default: "name").}
}
\value{
\code{as.tree} returns a tree object.
}
\description{
\code{as.tree} returns a tree object
}

