% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-tree.R
\docType{methods}
\name{distMatrix,tree-method}
\alias{distMatrix,tree-method}
\title{Compute a matrix of distances between all nodes.}
\usage{
\S4method{distMatrix}{tree}(object)
}
\arguments{
\item{object}{A tree object.}
}
\value{
\code{distMatrix} returns a matrix.
}
\description{
\code{distMatrix} returns a matrix of distances between all nodes.
}

