% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-tree.R
\docType{methods}
\name{distNodes,tree-method}
\alias{distNodes,tree-method}
\title{Compute distance between two nodes.}
\usage{
\S4method{distNodes}{tree}(object, node1.id, node2.id)
}
\arguments{
\item{object}{A tree object.}

\item{node1.id}{The ID of the 1st node.}

\item{node2.id}{The ID of the 2nd node.}
}
\value{
\code{distNodes} returns a numeric value.
}
\description{
\code{distNodes} returns the distance between two nodes.
}

