% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-tree.R
\docType{methods}
\name{distToRoot,tree-method}
\alias{distToRoot,tree-method}
\title{Compute distance to the Root node}
\usage{
\S4method{distToRoot}{tree}(object, node.id)
}
\arguments{
\item{object}{A tree object.}

\item{node.id}{The ID of the node.}
}
\value{
\code{distToRoot} returns a numeric value.
}
\description{
\code{distToRoot} returns the distance from a particular node to the root node.
}

