% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-tree.R
\docType{methods}
\name{getPathToRoot,tree-method}
\alias{getPathToRoot,tree-method}
\title{Get path to root.}
\usage{
\S4method{getPathToRoot}{tree}(object, node.id)
}
\arguments{
\item{object}{A tree object.}

\item{node.id}{The ID of the 1st node.}
}
\value{
\code{getPathToRoot} returns a character vector.
}
\description{
\code{getPathToRoot} returns the path from a particular node to root.
}

