% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-tree.R
\docType{methods}
\name{lca,tree-method}
\alias{lca,tree-method}
\title{Find the lowest common ancestor.}
\usage{
\S4method{lca}{tree}(object, node1.id, node2.id)
}
\arguments{
\item{object}{A tree object.}

\item{node1.id}{The ID of the 1st node.}

\item{node2.id}{The ID of the 2nd node.}
}
\value{
\code{lca} returns a character value.
}
\description{
\code{lca} returns the ID of the lowest common ancestor between two nodes.
}

