% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-tree.R
\docType{class}
\name{tree-class}
\alias{.tree}
\alias{tree-class}
\title{An S4 class to represent a tree.}
\description{
An S4 class to represent a tree.
}
\section{Slots}{

\describe{
\item{\code{tree.list}}{A list of nodes (node class)}
}}

